login-system/myproject/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.8
django-class-auth-views==1.0.2
django-phone-field==1.8.1
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
wincertstore==0.2
gunicorn